requirements.txt
Django==2.2.7
dj-database-url==0.5.0
django-cors-headers==3.2.0
django-debug-toolbar==2.1
django-querycount==0.7.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
drf-writable-nested==0.5.3
gunicorn==20.0.0
psycopg2-binary==2.8.4
python-dotenv==0.10.3
